fixed
Fixed issue with counting of push ups aswell as not counting when there are no previous runs of the program. Also it checks if there are previous runs with saved push ups if so counts them all and lets the program run if pushups bigger than 100 it says daily challange complete ofcource can be changed.